Updated Scenario. Added Simdis projects.

+++ SimulationInsight.ScenarioGenerator/Source/Flightpath/FlightpathData.cs
@@ -4,56 +4,56 @@
     {
         public int FlightpathId { get; init; }
 
-        public int Time { get; init; }
+        public double Time { get; init; }
 
-        public int LatitudeDeg { get; init; }
+        public double LatitudeDeg { get; init; }
 
-        public int LongitudeDeg { get; init; }
+        public double LongitudeDeg { get; init; }
 
-        public int Altitude { get; init; }
+        public double Altitude { get; init; }
 
-        public int PositionNorth { get; init; }
+        public double PositionNorth { get; init; }
 
-        public int PositionEast { get; init; }
+        public double PositionEast { get; init; }
 
-        public int PositionDown { get; init; }
+        public double PositionDown { get; init; }
 
-        public int VelocityNorth { get; init; }
+        public double VelocityNorth { get; init; }
 
-        public int VelocityEast { get; init; }
+        public double VelocityEast { get; init; }
 
-        public int VelocityDown { get; init; }
+        public double VelocityDown { get; init; }
 
-        public int TotalSpeed { get; init; }
+        public double TotalSpeed { get; init; }
 
-        public int TotalSpeed_kts { get; init; }
+        public double TotalSpeed_kts { get; init; }
 
-        public int GroundSpeed { get; init; }
+        public double GroundSpeed { get; init; }
 
-        public int GroundSpeed_kts { get; init; }
+        public double GroundSpeed_kts { get; init; }
 
-        public int AccelerationNorth { get; init; }
+        public double AccelerationNorth { get; init; }
 
-        public int AccelerationEast { get; init; }
+        public double AccelerationEast { get; init; }
 
-        public int AccelerationDown { get; init; }
+        public double AccelerationDown { get; init; }
 
-        public int AccelerationAxial { get; init; }
+        public double AccelerationAxial { get; init; }
 
-        public int AccelerationLateral { get; init; }
+        public double AccelerationLateral { get; init; }
 
-        public int AccelerationVetical { get; init; }
+        public double AccelerationVetical { get; init; }
 
-        public int HeadingAngleDeg { get; init; }
+        public double HeadingAngleDeg { get; init; }
 
-        public int PitchAngleDeg { get; init; }
+        public double PitchAngleDeg { get; init; }
 
-        public int BankAngleDeg { get; init; }
+        public double BankAngleDeg { get; init; }
 
-        public int HeadingRateDeg { get; init; }
+        public double HeadingRateDeg { get; init; }
 
-        public int PitchRateDeg { get; init; }
+        public double PitchRateDeg { get; init; }
 
-        public int BankRateDeg { get; init; }
+        public double BankRateDeg { get; init; }
     }
 }
